+++ lib/core/options/shape/vertex_shape.dart
@@ -7,13 +7,13 @@ import 'dart:ui';
 import 'package:flame/collisions.dart';
 import 'package:flutter_graph_view/flutter_graph_view.dart';
 
-/// Used to customize the vertex UI. Maybe the UI of the edge will also use this.
+/// Used to customize the vertex UI.
 ///
-/// 用于自定义节点的UI，可能边的定制也会用这个接口，待定。
-abstract class OptionShape {
-  /// render the shape by canvas.
+/// 用于自定义节点的UI。
+abstract class VertexShape {
+  /// render the vertex shape to canvas by data.
   ///
-  /// 通过Canvas绘制自定义的图形。
+  /// 通过节点数据将自定义的图形绘制到画布中。
   render(Vertex vertex, Canvas canvas, Paint paint, List<Paint> paintLayers);
 
   /// compute the width of the shape from data, used for overlap and mouse event judgment.
@@ -25,6 +25,13 @@ abstract class OptionShape {
   ///
   /// 通过数据计算形状的高，用于重叠与鼠标事件判断。
   double height(Vertex vertex);
+
+  /// compute the size of the shape from data, used for overlap and mouse event judgment.
+  ///
+  /// 通过数据计算形状的尺寸，用于重叠与鼠标事件判断。
+  Vector2 size(Vertex vertex) {
+    return Vector2(width(vertex), height(vertex));
+  }
 
   /// create a hit box, used for overlap judgment.
   ///
@@ -40,4 +47,15 @@ abstract class OptionShape {
   ///
   /// 根据数据更新画笔属性。
   void setPaint(Vertex vertex);
+
+  /// When some elements are activated and do not contain the current element.
+  ///
+  /// 当一些元素被激活且不包含当前元素
+  bool isWeaken(Vertex vertex) {
+    var cpn = vertex.cpn!;
+    var graph = cpn.gameRef.graph;
+    return graph.hoverVertex != null &&
+        (vertex != graph.hoverVertex &&
+            !graph.hoverVertex!.neighbors.contains(vertex));
+  }
 }