+++ lib/widgets/edge_component.dart
@@ -7,8 +7,6 @@ import 'package:flame/experimental.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_graph_view/flutter_graph_view.dart';
 
-import '../core/util.dart';
-
 ///
 /// Flame Component.
 /// Logic used to handle the connecting line between nodes (display & interaction)
@@ -16,11 +14,10 @@ import '../core/util.dart';
 /// 引擎组件
 /// 用于处理节点之间连接线的逻辑（展示+交互）
 ///
-class EdgeComponent extends RectangleComponent
+class EdgeComponent extends ShapeComponent
     with HasGameRef<GraphComponent>, TapCallbacks, Hoverable {
   late final Edge edge;
   late ValueNotifier<double> scaleNotifier;
-  double strokeWidth = 1;
   Graph graph;
   BuildContext context;
 
@@ -30,18 +27,11 @@ class EdgeComponent extends RectangleComponent
           anchor: Anchor.centerLeft,
         );
 
-  /// Compute the line length by two vertex.
-  /// 通过两个节点的坐标，计算线的长度。
-  double len() =>
-      edge.end == null || edge.start.cpn == null || edge.end!.cpn == null
-          ? 10
-          : Util.distance(edge.start.cpn!.position, edge.end!.cpn!.position);
+  EdgeShape get edgeShape => gameRef.options.edgeShape;
 
   @override
-  onLoad() {
-    super.onLoad();
-    size = Vector2(len().toDouble(), strokeWidth);
-  }
+  void render(Canvas canvas) =>
+      edgeShape.render(edge, canvas, paint, paintLayers);
 
   /// Update edge display by data of `edge`.
   /// 根据节点数据，对界面进行更新
@@ -49,32 +39,16 @@ class EdgeComponent extends RectangleComponent
   void update(double dt) {
     super.update(dt);
     position = edge.start.cpn!.position;
-    size = Vector2(len().toDouble(), strokeWidth);
-    if (graph.hoverVertex != null &&
-        !graph.hoverVertex!.neighborEdges.contains(edge)) {
-      paint = Paint()..color = Colors.white.withOpacity(0);
-    } else {
-      paint = Paint()..color = Colors.white;
-    }
-    Offset offset = Offset(
-      (edge.end == null
-              ? edge.start.cpn!.position.x
-              : edge.end!.cpn!.position.x) -
-          (edge.start.cpn!.position.x),
-      (edge.end == null
-              ? edge.start.cpn!.position.y
-              : edge.end!.cpn!.position.y) -
-          (edge.start.cpn!.position.y),
-    );
-
-    angle = offset.direction;
+    size = edgeShape.size(edge);
+    angle = edgeShape.angle(edge);
+    edgeShape.setPaint(edge);
   }
 
   /// Line will be wider after mouse enter.
   /// 对被鼠标浮入的线增加显视宽度
   @override
   bool onHoverEnter(PointerHoverInfo info) {
-    strokeWidth = 4;
+    paint.strokeWidth = 4;
     gameRef.graph.hoverEdge = edge;
     gameRef.overlays.add('edge');
     return true;
@@ -84,7 +58,7 @@ class EdgeComponent extends RectangleComponent
   /// 当鼠标浮出时，将线变回原来的宽度。
   @override
   bool onHoverLeave(PointerHoverInfo info) {
-    strokeWidth = 1;
+    paint.strokeWidth = 1;
     gameRef.graph.hoverEdge = null;
     Future.delayed(const Duration(milliseconds: 300), () {
       gameRef.overlays.remove('edge');